moved util to util file
