+++ src/applications/check-in/TODO.md
@@ -1,6 +1,5 @@
 # TODO
 
 - Add unit tests for
-  - create analytics object
   - make api
 - confirm structure
